requirements.txt
pandas==1.3.3
numpy==1.21.2
azure.storage.blob==12.8.1
boto3==1.21.21
pika==1.2.0
botocore==1.24.21
flask==2.0.2
python-dotenv
